Add .gitignore and remove generated files

+++ projects/ProyectoSistemaDeVentasConControlDeStock/ProyectoSistemaDeVentasConControlDeStock/DTOs/ProductosDTOs/ActualizarProductoDTO.cs
@@ -15,7 +15,7 @@ namespace ProyectoSistemaDeVentasConControlDeStock.DTOs.ProductosDTOs
 
         [Required]
         [Column("precio")]
-        public int precio { get; set; }
+        public decimal precio { get; set; }
 
         [Required]
         [Column("stock")]
